Refactor parser
