+++ include/ast.h
@@ -1,0 +1,5 @@
+#ifndef TAN_INCLUDE_AST_H_
+#define TAN_INCLUDE_AST_H_
+#include "src/ast/astnode.h"
+
+#endif /* TAN_INCLUDE_AST_H_ */
