+++ src/ast/astnode.cpp
@@ -1,4 +1,4 @@
-#include "parsedef.h"
+#include "src/ast/astnode.h"
 
 namespace tanlang {
